requirements.txt
paddlepaddle-gpu
paddle_ocr
opencv-python
numpy
Pillow
tqdm
requests
mss